requirements.txt
Package            Version
------------------ -----------
alembic            1.15.1
annotated-types    0.7.0
anyio              4.9.0
bidict             0.23.1
build              1.2.2.post1
certifi            2025.1.31
cffi               1.17.1
charset-normalizer 3.4.1
click              8.1.8
cryptography       44.0.2
distro             1.9.0
docutils           0.21.2
fastapi            0.115.11
fpdf               1.7.2
greenlet           3.1.1
gunicorn           23.0.0
h11                0.14.0
httpcore           1.0.7
httpx              0.28.1
id                 1.5.0
idna               3.10
jaraco.classes     3.4.0
jaraco.context     6.0.1
jaraco.functools   4.1.0
jeepney            0.9.0
Jinja2             3.1.6
keyring            25.6.0
lazy_loader        0.4
Mako               1.3.9
markdown-it-py     3.0.0
MarkupSafe         3.0.2
mdurl              0.1.2
more-itertools     10.6.0
nh3                0.2.21
numpy              2.2.4
packaging          24.2
pandas             2.2.3
pip                25.0.1
platformdirs       4.3.6
psutil             7.0.0
pycparser          2.22
pydantic           2.10.6
pydantic_core      2.27.2
Pygments           2.19.1
PyMySQL            1.1.1
pyproject_hooks    1.2.0
python-dateutil    2.9.0.post0
python-engineio    4.11.2
python-multipart   0.0.20
python-socketio    5.12.1
pytz               2025.1
PyYAML             6.0.2
readme_renderer    44.0
redis              5.2.1
reflex             0.7.3
reflex-hosting-cli 0.1.35
requests-toolbelt  1.0.0
rfc3986            2.0.0
rich               13.9.4
SecretStorage      3.3.3
setuptools         76.1.0
shellingham        1.5.4
simple-websocket   1.1.0
six                1.17.0
sniffio            1.3.1
SQLAlchemy         2.0.39
sqlmodel           0.0.24
starlette-admin    0.14.1
tabulate           0.9.0
tomlkit            0.13.2
twine              6.1.0
typer              0.15.2
typing_extensions  4.12.2
tzdata             2025.1
urllib3            2.3.0
uvicorn            0.34.0
wheel              0.45.1
wrapt              1.17.2
wsproto            1.2.0
